GeneralMultiResolutionRendererTest: clean up test.
